built-in/TensorFlow/Official/cv/image_classification/ShuffleNetV1-1.0x-group3_ID2129_for_TensorFlow/requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
audioread==2.1.8
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cloudpickle==1.5.0
contextlib2==0.6.0.post1
cupy==8.0.0b4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
-e git://github.com/NVIDIA/dllogger@26a0f8f1958de2c0c460925ff6102a4d2486d6cc#egg=DLLogger
entrypoints==0.3
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
graphsurgeon==0.4.5
grpcio==1.31.0
h5py==2.10.0
horovod @ file:///opt/tensorflow/horovod-source/dist/horovod-0.19.5.tar.gz
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
install==1.3.4
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.14.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyter-tensorboard==0.2.0
jupyterlab==1.2.14
jupyterlab-server==1.2.0
jupytext==1.5.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.5
kiwisolver==1.2.0
librosa==0.7.1
llvmlite==0.30.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
mpi4py==3.0.3
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
nltk==3.4.5
notebook==6.0.3
numba==0.46.0
numpy==1.17.3
nvidia-dali-cuda110==0.24.0
nvidia-dali-tf-plugin-cuda110==0.24.0
nvtx-plugins==0.1.8
opencv-python @ file:///home/wlx/opencv_python-4.4.0.42-cp36-cp36m-manylinux2014_x86_64.whl
opt-einsum==3.3.0
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
portalocker==2.0.0
portpicker==1.3.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
psutil==5.7.0
ptyprocess==0.6.0
pycocotools==2.0.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-speech-features==0.6
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
resampy==0.2.2
sacrebleu==1.4.2
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.2.1
Send2Trash==1.5.0
sentencepiece==0.1.83
Shapely==1.7.1
six==1.13.0
SoundFile==0.10.3.post1
tensorboard @ file:///nvidia/opt/tensorboard_install/tensorboard-1.15.0%2Bnv-py3-none-any.whl
tensorboard-plugin-dlprof @ file:///nvidia/opt/tensorboard_install/tensorboard_plugin_dlprof-0.6-py3-none-any.whl
tensorflow @ file:///tmp/pip/tensorflow-1.15.3%2Bnv-cp36-cp36m-linux_x86_64.whl
tensorflow-estimator==1.15.1
tensorrt==7.1.3.4
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.1
torch==1.4.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typing==3.7.4.3
uff==0.6.9
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
